TechNews/requirements.txt
django
djangorestframework
scrapy
selenium
